+++ Editor/Variant Set Editor/TransformVariantSetEditor.cs
@@ -1,9 +1,13 @@
+using IndustryCSE.Tool.ProductConfigurator.ScriptableObjects;
 using UnityEditor;
+using UnityEngine;
+using System;
 
 namespace IndustryCSE.Tool.ProductConfigurator.Editor
 {
     [CustomEditor(typeof(TransformVariantSet))]
     public class TransformVariantSetEditor : CustomConfigurationEditorBase
     {
+        
     }
 }