+++ Runtime/Variant Set/GameObjectVariantSet.cs
@@ -47,9 +47,67 @@ namespace IndustryCSE.Tool.ProductConfigurator
         }
         
 #if UNITY_EDITOR
+        /// <summary>
+        /// Create a new variant with the given name and object
+        /// </summary>
+        /// <param name="variantName">Variant Name</param>
+        /// <param name="variantObject">Assign GameObject</param>
+        /// <typeparam name="T"></typeparam>
+        /// <returns></returns>
+        /// <exception cref="ArgumentException"></exception>
+        public override VariantAsset CreateVariantAsset<T>(string variantName, T variantObject)
+        {
+            if (variantObject == null)
+            {
+                throw new ArgumentException("variantObject cannot be null");
+            }
+            
+            if (!(variantObject is GameObject))
+            {
+                throw new ArgumentException("variantObject must be a GameObject");
+            }
+            
+            var variantAsset = CreateVariantAsset(variantName);
+            AssignVariantObject(variantAsset.UniqueIdString, variantObject as GameObject);
+            return variantAsset;
+        }
+
         public override void AddVariant(VariantAsset variantAsset)
         {
             variants.Add(new GameObjectVariant {variantAsset = variantAsset});
+        }
+
+        /// <summary>
+        /// Add a new variant with the given name and object
+        /// </summary>
+        /// <param name="variantName">Variant Name</param>
+        /// <param name="variantObject">Assign GameObject</param>
+        /// <typeparam name="T"></typeparam>
+        /// <returns></returns>
+        /// <exception cref="ArgumentException"></exception>
+        public override void AddVariant<T>(VariantAsset variantAsset, T variantObject)
+        {
+            if (variantObject == null)
+            {
+                throw new ArgumentException("variantObject cannot be null");
+            }
+            
+            if (!(variantObject is GameObject))
+            {
+                throw new ArgumentException("variantObject must be a GameObject");
+            }
+
+            AddVariant(variantAsset);
+            AssignVariantObject(variantAsset.UniqueIdString, variantObject as GameObject);
+        }
+        
+        private void AssignVariantObject(string id, GameObject targetGameObject)
+        {
+            var targetVariant = VariantBase.Find(x => string.Equals(x.variantAsset.UniqueIdString, id));
+            if (targetVariant != null)
+            {
+                ((GameObjectVariant)targetVariant).VariantGameObject = targetGameObject;
+            }
         }
 #endif
     }